report type save initial implementation
