+++ src/app/report-types/define-relationships/[reportTypeId]/page.tsx
@@ -1,0 +1,6 @@
+"use client";
+import DefineRelationshipsForm from "@/components/report-type/define-relationship-form";
+
+export default function SelectRelationshipPage({params}: {params: any}) {
+  return <DefineRelationshipsForm reportTypeId={params.reportTypeId} ></DefineRelationshipsForm>
+}
