+++ src/app/report-types/select-object/[reportTypeId]/page.tsx
@@ -1,0 +1,6 @@
+import SelectObjectForm from "@/components/report-type/select-object-form";
+
+export default function ReportTypeSelectionPage({ params }: { params: any }) {
+    console.log("Edit report type page ------------------------");
+    return <SelectObjectForm reportTypeId={params?.reportTypeId} ></SelectObjectForm>;
+}
