+++ src/app/report-types/select-object/page.tsx
@@ -1,8 +1,6 @@
 import SelectObjectForm from "@/components/report-type/select-object-form";
 
-
 export default function ReportTypeSelectionPage() {
-
-
-  return <SelectObjectForm></SelectObjectForm>
+  console.log("Create report type page ------------------------");
+  return <SelectObjectForm reportTypeId="" ></SelectObjectForm>
 }
